[Bioinfo-practical2018] prototype implementation for dual hgp workflow
This is a very basic implementation of dual hgp partitioning.
The algorithm first converts the input hypergraph into its dual.
The dual hypergraph is then stripped of all parallel nets and
subsequently partitioned.
After partitioning, the partition of the dual hypergraph is used to
infer a partition of the input hypergraph.
Hypernodes which could not be assigned to a block, because the
corresponding net in the dual hypergraph was cut, remain in block
'-1'. All other hypernodes are assigned the blocks according to the
partition of the dual hypergraph.

This version write the following files to disk (see the corresponding
console output):
1.) The dual hypergraph containing parallel hyperedges
2.) The partition of the dual hypergraph
3.) The partition of the input hypergraph, where 'undecided'
hypernodes are assigned to block '-1'.